requirements.txt

streamlit
speechrecognition
googletrans==4.0.0-rc1
python-docx
fpdf
pyaudio
